feat(1.2.1): Rename `classes` to `utils`
