+++ src/app/commons/icon/files/crews/pilote-de-chars.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const piloteDeCharsIcon: Icon = new Icon(
     'pilote_de_char',
