+++ src/app/commons/icon/files/crews/tireur.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const tireurIcon: Icon = new Icon(
     'tireur',
