+++ src/app/commons/icon/files/tanks/amx_m4_mle54.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const amxM4Mle54Icon: Icon = new Icon(
     'amx_m4_mle54',
