+++ src/app/commons/icon/files/tanks/cs-63.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const cs63Icon: Icon = new Icon(
     'cs-63',
