+++ src/app/commons/icon/files/tanks/is-7.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const is7Icon: Icon = new Icon(
     'is-7',
