+++ src/app/commons/icon/files/tanks/kampfpanzer_07_p(e).ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const kpz07pEIcon: Icon = new Icon(
     'kampfpanzer_07_p(e)',
