+++ src/app/commons/icon/files/tanks/leopard_1.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const leopard1Icon: Icon = new Icon(
     'leopard_1',
