+++ src/app/commons/icon/files/tanks/object_260.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const object260Icon: Icon = new Icon(
     'object_260',
