+++ src/app/commons/icon/files/tanks/object_279_early.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const object279EarlyIcon: Icon = new Icon(
     'object_279_early',
