+++ src/app/commons/icon/files/tanks/strv_103b.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const strv103bIcon: Icon = new Icon(
     'strv_103b',
