+++ src/app/commons/icon/files/tanks/super_conqueror.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const superConquerorIcon: Icon = new Icon(
     'super_conqueror',
