+++ src/app/commons/icon/files/tanks/t110e3.icon.ts
@@ -1,4 +1,4 @@
-import { Icon } from 'src/app/commons/classes/icon.class';
+import { Icon } from 'src/app/commons/utils/icon.util';
 
 export const t110e3Icon: Icon = new Icon(
     't110e3',
