+++ src/app/commons/utils/svg-custom.util.ts
@@ -1,6 +1,6 @@
 export class SvgCustom {
-    public static sun =
+    public static sun: string =
         'M11.6625 0.0879108C11.5547 0.144186 11.4469 0.252045 11.3906 0.364595C11.3016 0.533419 11.2969 0.63659 11.2969 1.82774C11.2969 2.98606 11.3063 3.12675 11.3859 3.27212C11.5219 3.53005 11.6859 3.62384 12 3.62384C12.3141 3.62384 12.4781 3.53005 12.6141 3.27212C12.6938 3.12675 12.7031 2.98606 12.7031 1.82774C12.7031 0.63659 12.6984 0.533419 12.6094 0.364595C12.4359 0.022257 12.0141 -0.0996716 11.6625 0.0879108Z M3.72654 3.36594C3.67497 3.3847 3.57185 3.46911 3.49685 3.54883C3.30935 3.74579 3.27185 4.13034 3.41247 4.36012C3.46872 4.44454 3.89529 4.89942 4.36872 5.36838C5.06716 6.06244 5.25935 6.22657 5.41404 6.26878C5.94372 6.40946 6.41716 5.93113 6.27654 5.4059C6.23435 5.25583 6.0656 5.05418 5.42341 4.40233C4.43435 3.40345 4.32185 3.31904 4.03591 3.32373C3.91872 3.32373 3.7781 3.34249 3.72654 3.36594Z M19.5656 3.45504C19.4671 3.53007 19.0218 3.95682 18.5765 4.40233C17.9249 5.06356 17.7656 5.25583 17.7234 5.4059C17.5827 5.93582 18.0609 6.40947 18.5859 6.26878C18.7359 6.22657 18.9421 6.05306 19.6312 5.36838C20.1046 4.89943 20.5312 4.44454 20.5874 4.36013C20.7046 4.16785 20.6906 3.77393 20.5499 3.58635C20.3343 3.29091 19.8702 3.22994 19.5656 3.45504Z M10.9454 6.10931C8.80786 6.50792 7.02192 8.05548 6.32817 10.1189C5.95786 11.2209 5.95317 12.7122 6.31411 13.8189C6.99848 15.9246 8.82661 17.5096 10.9875 17.8801C11.5313 17.9739 12.4782 17.9692 13.0313 17.8754C15.1969 17.5049 17.0016 15.9246 17.6907 13.8002C18.0469 12.7028 18.0469 11.2866 17.6907 10.1892C17.1375 8.47754 15.8766 7.12694 14.2032 6.44227C13.5329 6.16558 13.0735 6.08117 12.1641 6.05772C11.55 6.04834 11.2172 6.06241 10.9454 6.10931Z M0.365625 11.3898C0.164063 11.4929 0 11.7696 0 11.9947C0 12.2292 0.164063 12.4965 0.375 12.6044C0.54375 12.6935 0.646875 12.6982 1.8375 12.6982C3.24375 12.6982 3.32344 12.6841 3.53906 12.398C3.67031 12.2198 3.67031 11.7696 3.53906 11.5914C3.32344 11.3054 3.24844 11.2913 1.82813 11.2913C0.6375 11.296 0.525 11.3007 0.365625 11.3898Z M20.7188 11.3804C20.4609 11.5164 20.3672 11.6805 20.3672 11.9947C20.3672 12.3089 20.4609 12.4731 20.7188 12.609C20.8641 12.6888 21.0047 12.6982 22.1625 12.6982C23.3531 12.6982 23.4563 12.6935 23.625 12.6044C23.8359 12.4965 24 12.2292 24 11.9947C24 11.7602 23.8359 11.4929 23.625 11.3851C23.4563 11.296 23.3531 11.2913 22.1625 11.2913C21.0047 11.2913 20.8641 11.3007 20.7188 11.3804Z M5.37187 17.7394C5.23593 17.7863 4.95468 18.0349 4.34999 18.6398C3.88593 19.0994 3.46874 19.5449 3.41249 19.6293C3.33749 19.7513 3.32343 19.8357 3.33749 20.042C3.35624 20.2624 3.38436 20.3281 3.52968 20.4688C3.6703 20.6141 3.73593 20.6423 3.95624 20.661C4.1578 20.6751 4.24687 20.661 4.36874 20.586C4.45312 20.5297 4.9078 20.103 5.37655 19.6293C6.0703 18.9259 6.23437 18.7383 6.27655 18.5789C6.42187 18.0302 5.90624 17.5472 5.37187 17.7394Z M18.1078 17.7582C17.7844 17.9083 17.6297 18.2318 17.7187 18.5695C17.7656 18.7383 17.9015 18.8978 18.5765 19.5871C19.5562 20.5766 19.6781 20.6704 19.9547 20.6704C20.3765 20.6704 20.6297 20.4453 20.6625 20.0326C20.6765 19.8404 20.6625 19.7513 20.5875 19.6293C20.5312 19.5449 20.1047 19.09 19.6312 18.6211C18.9375 17.9364 18.7359 17.7629 18.5859 17.7207C18.3422 17.655 18.3422 17.655 18.1078 17.7582Z M11.6906 20.4078C11.6109 20.4406 11.4937 20.5438 11.4328 20.6282C11.3203 20.7876 11.3203 20.8017 11.3062 22.1101C11.2922 23.3669 11.2969 23.4419 11.3906 23.6201C11.4984 23.8359 11.7656 24 12 24C12.2344 24 12.5015 23.8359 12.6094 23.6248C12.6984 23.456 12.7031 23.3528 12.7031 22.1617C12.7031 21.0034 12.6937 20.8627 12.614 20.7173C12.5672 20.6282 12.4734 20.5157 12.4078 20.4641C12.2531 20.3515 11.8828 20.3187 11.6906 20.4078Z';
-    public static moon =
+    public static moon: string =
         'M11.6599 0.0581877C11.44 0.154776 11.2146 0.428443 11.1717 0.64845C11.1556 0.750404 11.1502 1.09919 11.161 1.42116C11.1717 1.9202 11.1932 2.03825 11.2897 2.16167C11.499 2.44607 11.6707 2.53192 12.0087 2.53192C12.3467 2.53192 12.5184 2.44607 12.7276 2.16167C12.8296 2.02752 12.8403 1.93629 12.8403 1.27091C12.8403 0.460639 12.8135 0.374783 12.4808 0.127946C12.3038 -0.00620449 11.8746 -0.0437666 11.6599 0.0581877Z M3.77319 3.33145C3.55322 3.42804 3.32788 3.7017 3.28496 3.92708C3.21521 4.32416 3.29032 4.48514 3.81611 5.01101C4.34727 5.53688 4.50286 5.61737 4.91062 5.53688C5.17351 5.48859 5.49543 5.16663 5.54371 4.90369C5.61883 4.51197 5.53835 4.34026 5.07157 3.85732C4.4975 3.26169 4.20241 3.14364 3.77319 3.33145Z M19.5737 3.31536C19.4074 3.37975 18.6294 4.14173 18.5221 4.34027C18.4095 4.54954 18.4255 4.96809 18.5543 5.17737C18.7045 5.4242 18.9621 5.56372 19.2679 5.56372C19.6166 5.56372 19.7454 5.4886 20.2551 4.96273C20.7219 4.48515 20.8023 4.30807 20.7326 3.92708C20.6414 3.43878 20.0512 3.12755 19.5737 3.31536Z M9.46029 4.45294C8.34433 4.87149 7.46443 5.43492 6.61673 6.28811C5.49003 7.41498 4.77109 8.73502 4.43309 10.3019C4.2614 11.0961 4.2614 12.5986 4.43309 13.382C5.1252 16.5748 7.44297 18.8929 10.6353 19.5797C11.4186 19.7461 12.9262 19.7461 13.7095 19.5797C16.4619 18.9841 18.6509 17.0845 19.5737 14.4981C19.6971 14.144 19.7347 13.9562 19.7186 13.7898C19.6864 13.4786 19.4771 13.2156 19.166 13.0976C18.9245 13.0064 18.9031 13.0064 18.4363 13.1191C16.7677 13.5269 15.2333 13.3283 13.79 12.5127C12.6365 11.8634 11.6171 10.699 11.1342 9.48089C10.6299 8.20378 10.5494 6.9696 10.8874 5.5798C11.0055 5.10759 11.0055 5.08613 10.9143 4.84466C10.6997 4.28122 10.249 4.16317 9.46029 4.45294Z M0.500367 11.2195C-0.0200583 11.4502 -0.164919 12.1371 0.210646 12.5664C0.435985 12.8186 0.661324 12.8722 1.39636 12.8508C2.07237 12.8347 2.18504 12.7917 2.42648 12.4644C2.58207 12.2551 2.58207 11.7507 2.42648 11.5414C2.17968 11.2087 2.07774 11.1712 1.34271 11.1551C0.854471 11.1497 0.613037 11.1658 0.500367 11.2195Z M21.9612 11.2195C21.4408 11.4502 21.2959 12.1371 21.6715 12.5664C21.8968 12.8186 22.1221 12.8722 22.8572 12.8508C23.3883 12.84 23.5064 12.8186 23.6298 12.722C23.9141 12.5127 24 12.341 24 12.0029C24 11.6649 23.9141 11.4931 23.6298 11.2839C23.501 11.1873 23.3937 11.1712 22.8035 11.1551C22.3153 11.1497 22.0739 11.1658 21.9612 11.2195Z M4.30962 18.5441C4.22378 18.5978 3.96088 18.8339 3.73554 19.07C3.2956 19.5261 3.21512 19.7032 3.29023 20.0895C3.33852 20.3525 3.66043 20.6744 3.92333 20.7227C4.31499 20.7978 4.48667 20.7173 4.96954 20.2505C5.4846 19.7461 5.57045 19.6119 5.57045 19.2953C5.57045 18.6622 4.83005 18.2221 4.30962 18.5441Z M18.9568 18.5173C18.7368 18.6139 18.5115 18.8875 18.4686 19.1129C18.3988 19.51 18.4739 19.671 18.9997 20.1968C19.5309 20.7227 19.6865 20.8032 20.0942 20.7227C20.3571 20.6744 20.679 20.3524 20.7273 20.0895C20.8024 19.6978 20.7219 19.5261 20.2552 19.0431C19.6811 18.4475 19.386 18.3295 18.9568 18.5173Z M11.6599 21.5222C11.44 21.6188 11.2146 21.8925 11.1717 22.1125C11.1556 22.2145 11.1502 22.5632 11.161 22.8852C11.1771 23.5291 11.2254 23.6472 11.5473 23.8833C11.7565 24.0389 12.2608 24.0389 12.4701 23.8833C12.8135 23.6311 12.8403 23.5452 12.8403 22.735C12.8403 21.9247 12.8135 21.8388 12.4808 21.592C12.3038 21.4578 11.8746 21.4203 11.6599 21.5222Z';
 }
