+++ src/app/commons/utils/windows-custom.util.ts
